Introduce a GroupContext
This groups the functionalities around groups into a single class.
