+++ tests/behat/behat.yml
@@ -19,6 +19,7 @@ default:
         - Drupal\DrupalExtension\Context\ConfigContext
         - Drupal\social\Behat\CKEditorContext
         - Drupal\social\Behat\EmailContext
+        - Drupal\social\Behat\GroupContext
         - Drupal\social\Behat\PostContext
         - Drupal\social\Behat\FeatureContext
         - Drupal\social\Behat\SocialDrupalContext
